+++ Accueil.js
@@ -1,10 +1,9 @@
-import React, { useState } from 'react'
-import Article from './Article'
-import { imageResizer, accessibleImage } from './Article'
-import { Switch, Route, BrowserRouter as Router, Link } from 'react-router-dom'
+import React from 'react'
+import { BrowserRouter as Router, Link, Route, Switch } from 'react-router-dom'
+import Article, { accessibleImage, imageResizer } from './Article'
+import Projects from './Projects'
 import ScrollToTop from './ScrollToTop'
-import projets from './projets.yaml'
-import styled from 'styled-components'
+import SubHeader from './SubHeader'
 
 export const dateCool = (date) =>
 	new Date(date).toLocaleString(undefined, {
@@ -139,130 +138,6 @@ const BlogHeader = () => (
 		`}
 	></div>
 )
-
-const SubHeader = styled.header`
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	h2 {
-		font-size: 120%;
-		font-weight: normal;
-		max-width: 20rem;
-		/*
-					border-bottom: 2px solid #fe9d00;
-					padding: 0.2rem 0.4rem;
-					color: #38255c;
-					*/
-	}
-
-	h2 em {
-		background: rgb(78, 47, 90);
-		font-style: normal;
-		color: white;
-		padding: 0rem 0.6rem 0.1rem;
-	}
-	> img {
-		width: 2rem;
-		margin: 0 0.4rem;
-		transform: scale(-1, 1) translateY(0.5rem);
-	}
-`
-
-const Projects = () => {
-	const [descriptionShown, showDescription] = useState(false)
-
-	const selectedProject = projets.find(({ nom }) => nom === descriptionShown)
-
-	return (
-		<div
-			css={`
-				margin: 0 0 2rem;
-			`}
-		>
-			<SubHeader>
-				<img src="https://openmoji.org/data/color/svg/2935.svg" />
-				<h2>
-					<em>Mes projets</em>
-				</h2>
-			</SubHeader>
-			<ul
-				css={`
-					display: flex;
-					flex-wrap: wrap;
-					list-style-type: none;
-					justify-content: center;
-					align-items: center;
-					padding-left: 0;
-					flex-wrap: nowrap;
-					overflow-x: auto;
-					justify-content: normal;
-					height: 11rem;
-					background: linear-gradient(0% #4d2d5b, 100% #7b3657);
-					@media (max-width: 800px) {
-						scrollbar-width: none;
-					}
-					li {
-						padding: 0.1rem 0rem;
-						margin: 0.4rem;
-						border-radius: 0.2rem;
-						line-height: 1.6rem;
-						display: flex;
-						flex-direction: column;
-						justify-content: center;
-						align-items: center;
-						box-shadow: 0.5px 1px 1px hsl(var(--shadow-color) / 0.7);
-						--shadow-color: 0deg 0% 50%;
-						min-width: 10rem;
-						height: 8rem;
-						text-align: center;
-					}
-					li img {
-						height: 4rem;
-						max-width: 100%;
-						transition: filter 0.3s ease-in;
-					}
-					li img:hover,
-					li img:active {
-					}
-					li h3 {
-						font-size: 100%;
-						margin: 0.4rem;
-					}
-				`}
-			>
-				{projets.map(({ lien, nom, image, github }) => (
-					<button
-						onClick={() =>
-							showDescription(selectedProject?.nom === nom ? false : nom)
-						}
-					>
-						<li css={``} key={lien}>
-							<img src={image} />
-							<h3>{nom}</h3>
-						</li>
-					</button>
-				))}
-			</ul>
-			{descriptionShown && (
-				<div
-					css={`
-						background: #4e2f5a;
-						color: white;
-						padding: 0.6rem 1rem;
-						p {
-							margin: 0 0 0.6rem;
-						}
-					`}
-				>
-					<p>{selectedProject.description}</p>
-					<a href={selectedProject.lien}>
-						{selectedProject.lien.replace('https://', '')}
-					</a>
-				</div>
-			)}
-		</div>
-	)
-}
 
 let Liste = ({ articles }) => (
 	<main css={pageLayout}>
