+++ package.json
@@ -13,6 +13,7 @@
 		"@babel/plugin-transform-react-jsx": "^7.12.13",
 		"buffer": "^6.0.2",
 		"copy-webpack-plugin": "^10.2.0",
+		"downa": "^1.0.1",
 		"front-matter-loader": "^2.0.0",
 		"html-loader": "^1.3.2",
 		"imports-loader": "^1.2.0",
